ci: Build windows wheel with OSGeo4W
This introduces an initial CI with a `build` stage. This stage
contains a job to build a windows wheel based on the python version
from OSGeo4W. It is based on a docker image from SFCGAL registry which
contains a compiled SFCGAL and its dependencies installed with vcpkg.

This job has the following rules:
 - optional when a new merge request is created
 - automatic on a push to the master branch or when a new tag is
 created.