refactor: wrap_geom becomes a method of the Geometry class
We can't call wrap_geom several times in a loop, because of memory
issue (same C geometry across different Python Geometry
instances). Then we clone the C geometry in the Geometry.wrap method.

And we keep the previous wrap_geom as a Geometry.from_c_geometry class
method. In order to consider unsupported geometry classes (see #5 for
an illustration), an if-clause prevents the KeyError.

fix #24